Implementation of day 3 part 2 works on real data

+++ aoc-2020/src/main/kotlin/net/bbelovic/adventofcode/year2020/day3/Day3PuzzlePart1.kt
@@ -1,7 +1,7 @@
 package net.bbelovic.adventofcode.year2020.day3
 
 class Day3PuzzlePart1 {
-    fun solve(input: List<String>): Int {
+    fun solve(input: List<String>): Long {
         val tobogganRacer = TobogganRacer(listOf(Pair(3, 1)))
         return tobogganRacer.solve(input)
