+++ aoc-2020/src/main/kotlin/net/bbelovic/adventofcode/year2020/day3/Day3PuzzlePart2.kt
@@ -2,7 +2,7 @@ package net.bbelovic.adventofcode.year2020.day3
 
 class Day3PuzzlePart2 {
 
-    fun solve(input: List<String>): Int {
+    fun solve(input: List<String>): Long {
         val slopes = listOf(Pair(1, 1), Pair(3, 1), Pair(5, 1), Pair(7, 1), Pair(1, 2))
         val tobogganRacer = TobogganRacer(slopes)
         return tobogganRacer.solve(input)
